requirements.txt
streamlit
google-generativeai
requests
beautifulsoup4
matplotlib
pandas
numpy
pillow
googlesearch-python
